i promise i am behaving and rearranging the long ass spaghetti code
