+++ pcs/src/lib.rs
@@ -5,5 +5,3 @@ pub mod raw;
 
 pub mod orion;
 pub use self::orion::*;
-#[cfg(test)]
-mod orion_test;
