+++ pcs/src/orion/tests.rs
@@ -8,12 +8,13 @@ use mersenne31::{M31Ext3, M31Ext3x16, M31x16, M31};
 use polynomials::{EqPolynomial, MultiLinearPoly};
 use transcript::{BytesHashTranscript, Keccak256hasher, Transcript};
 
-use crate::{
-    simd_inner_prod, transpose_in_place, OrionCode, OrionCommitment, ORION_CODE_PARAMETER_INSTANCE,
-    ORION_PCS_SOUNDNESS_BITS,
+use crate::orion::{
+    orion_code::{OrionCode, ORION_CODE_PARAMETER_INSTANCE},
+    pcs_impl::{
+        OrionCommitment, OrionCommitmentWithData, OrionPublicParams, ORION_PCS_SOUNDNESS_BITS,
+    },
+    utils::{simd_inner_prod, transpose_in_place},
 };
-
-use super::{OrionCommitmentWithData, OrionPublicParams};
 
 fn column_combination<F, PackF>(mat: &[F], combination: &[F]) -> Vec<F>
 where
@@ -91,39 +92,37 @@ fn test_orion_code() {
     });
 }
 
-impl OrionPublicParams {
-    fn dumb_commit<F, ComPackF>(
-        &self,
-        poly: &MultiLinearPoly<F>,
-    ) -> OrionCommitmentWithData<F, ComPackF>
-    where
-        F: Field + FieldSerde,
-        ComPackF: SimdField<Scalar = F>,
-    {
-        let (row_num, msg_size) = Self::row_col_from_variables::<F>(poly.get_num_vars());
+fn dumb_commit<F, ComPackF>(
+    orion_pp: &OrionPublicParams,
+    poly: &MultiLinearPoly<F>,
+) -> OrionCommitmentWithData<F, ComPackF>
+where
+    F: Field + FieldSerde,
+    ComPackF: SimdField<Scalar = F>,
+{
+    let (row_num, msg_size) = OrionPublicParams::row_col_from_variables::<F>(poly.get_num_vars());
 
-        let mut interleaved_codewords: Vec<_> = poly
-            .coeffs
-            .chunks(msg_size)
-            .flat_map(|msg| self.code_instance.encode(&msg).unwrap())
-            .collect();
+    let mut interleaved_codewords: Vec<_> = poly
+        .coeffs
+        .chunks(msg_size)
+        .flat_map(|msg| orion_pp.code_instance.encode(&msg).unwrap())
+        .collect();
 
-        let mut scratch = vec![F::ZERO; row_num * self.code_len()];
-        transpose_in_place(&mut interleaved_codewords, &mut scratch, row_num);
-        drop(scratch);
+    let mut scratch = vec![F::ZERO; row_num * orion_pp.code_len()];
+    transpose_in_place(&mut interleaved_codewords, &mut scratch, row_num);
+    drop(scratch);
 
-        if !interleaved_codewords.len().is_power_of_two() {
-            let aligned_po2_len = interleaved_codewords.len().next_power_of_two();
-            interleaved_codewords.resize(aligned_po2_len, F::default());
-        }
+    if !interleaved_codewords.len().is_power_of_two() {
+        let aligned_po2_len = interleaved_codewords.len().next_power_of_two();
+        interleaved_codewords.resize(aligned_po2_len, F::default());
+    }
 
-        let interleaved_alphabet_tree =
-            tree::Tree::compact_new_with_field_elems::<F, ComPackF>(&interleaved_codewords);
+    let interleaved_alphabet_tree =
+        tree::Tree::compact_new_with_field_elems::<F, ComPackF>(&interleaved_codewords);
 
-        OrionCommitmentWithData {
-            interleaved_alphabet_tree,
-            _phantom: PhantomData,
-        }
+    OrionCommitmentWithData {
+        interleaved_alphabet_tree,
+        _phantom: PhantomData,
     }
 }
 
@@ -139,7 +138,7 @@ where
         OrionPublicParams::from_random::<F>(num_vars, ORION_CODE_PARAMETER_INSTANCE, &mut rng);
 
     let real_commit = orion_pcs.commit::<F, ComPackF>(&random_poly).unwrap();
-    let dumb_commit = orion_pcs.dumb_commit::<F, ComPackF>(&random_poly);
+    let dumb_commit = dumb_commit::<F, ComPackF>(&orion_pcs, &random_poly);
 
     let real_commitment: OrionCommitment = real_commit.into();
     let dumb_commitment: OrionCommitment = dumb_commit.into();
@@ -291,68 +290,5 @@ fn test_orion_pcs_open() {
     });
     (9..=15).for_each(|num_vars| {
         test_orion_pcs_open_generics::<M31, M31Ext3, M31x16, M31x16, M31Ext3x16>(num_vars)
-    })
-}
-
-fn test_orion_pcs_full_e2e_generics<F, EvalF, ComPackF, IPPackF, IPPackEvalF>(num_vars: usize)
-where
-    F: Field + FieldSerde,
-    EvalF: Field + FieldSerde + Mul<F, Output = EvalF> + From<F>,
-    ComPackF: SimdField<Scalar = F>,
-    IPPackF: SimdField<Scalar = F>,
-    IPPackEvalF: SimdField<Scalar = EvalF> + Mul<IPPackF, Output = IPPackEvalF>,
-{
-    let mut rng = test_rng();
-
-    let random_poly = MultiLinearPoly::<F>::random(num_vars, &mut rng);
-    let random_poly_ext = MultiLinearPoly::new(
-        random_poly
-            .coeffs
-            .iter()
-            .cloned()
-            .map(EvalF::from)
-            .collect(),
-    );
-    let random_point: Vec<_> = (0..num_vars)
-        .map(|_| EvalF::random_unsafe(&mut rng))
-        .collect();
-    let expected_eval = random_poly_ext.evaluate_jolt(&random_point);
-
-    let mut transcript: BytesHashTranscript<EvalF, Keccak256hasher> = BytesHashTranscript::new();
-    let mut transcript_cloned = transcript.clone();
-
-    let orion_pp =
-        OrionPublicParams::from_random::<F>(num_vars, ORION_CODE_PARAMETER_INSTANCE, &mut rng);
-
-    let commit_with_data = orion_pp.commit::<F, ComPackF>(&random_poly).unwrap();
-
-    let (_, opening) = orion_pp.open::<F, EvalF, ComPackF, IPPackF, IPPackEvalF, _>(
-        &random_poly,
-        &commit_with_data,
-        &random_point,
-        &mut transcript,
-    );
-
-    assert!(
-        orion_pp.verify::<F, ComPackF, EvalF, IPPackF, IPPackEvalF, _>(
-            &commit_with_data.into(),
-            &random_point,
-            expected_eval,
-            &opening,
-            &mut transcript_cloned
-        )
-    );
-}
-
-#[test]
-fn test_orion_pcs_full_e2e() {
-    (19..=25).for_each(|num_vars| {
-        test_orion_pcs_full_e2e_generics::<GF2, GF2_128, GF2x8, GF2x8, GF2_128x8>(num_vars);
-        test_orion_pcs_full_e2e_generics::<GF2, GF2_128, GF2x64, GF2x8, GF2_128x8>(num_vars);
-        test_orion_pcs_full_e2e_generics::<GF2, GF2_128, GF2x128, GF2x8, GF2_128x8>(num_vars);
-    });
-    (9..=15).for_each(|num_vars| {
-        test_orion_pcs_full_e2e_generics::<M31, M31, M31x16, M31x16, M31x16>(num_vars);
-        test_orion_pcs_full_e2e_generics::<M31, M31Ext3, M31x16, M31x16, M31Ext3x16>(num_vars);
     })
 }
